Fixed events format and added more info to the events :)

+++ handlers/api/DeleteBucket.go
@@ -12,6 +12,13 @@ import (
 // If no bucket is found (or the bucket name is incorrect) it throws an error.
 // AWS Documentation: https://docs.aws.amazon.com/AmazonS3/latest/API/API_DeleteBucket.html
 func (h *Handler) DeleteBucket(c *gin.Context) {
+	clientIp := utils.GetClientIP(c.Request)
+	accessKey := c.GetString("accessKey")
+	if accessKey == "" {
+		utils.HandleError(c, utils.InternalServerError, "Access key not found")
+		return
+	}
+
 	bucket := utils.ClearInput(c.Param("bucket"))
 	ok, e := h.Storage.DeleteBucket(bucket)
 	if !ok {
@@ -19,7 +26,7 @@ func (h *Handler) DeleteBucket(c *gin.Context) {
 		return
 	}
 
-	go events.HandleEventBucket(h.Store, types.EventBucketDeleted, bucket, "")
+	go events.HandleEventBucket(h.Store, types.EventBucketDeleted, bucket, accessKey, clientIp)
 
 	c.Status(http.StatusNoContent)
 }
